Fix mismarking of automatically installed packages
When the option 'markauto' is set, packages installed to satisfy
dependencies (as opposed to those explicitly listed in packages=
lines) should be marked as being automatically installed (see
apt-mark(8)).

There was an error in the comparison of package names such that
any package of which the name was a substring of the name of an
explicitly listed package was mistakenly considered a manually
installed package and not marked as being automatically
installed.